requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# AI Provider SDKs
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.8.1
groq==0.4.1
deepl==1.18.0
azure-ai-translation-text==1.0.0

# Utilities
requests==2.31.0
httpx==0.26.0
langdetect==1.0.9
tiktoken==0.5.2

# Export Libraries
reportlab==4.0.9
ebooklib==0.18
python-docx==1.1.0

# Additional Features
pandas==2.1.4
openpyxl==3.1.2
schedule==1.2.0

